requirements.txt
pytest

pytest-selenium

urllib.parse

urllib3

pytest-base-url